GraDos/marker-worker/pyproject.toml
[project]
name = "grados-marker-worker"
version = "0.1.0"
requires-python = ">=3.12,<3.13"
dependencies = [
  "marker-pdf==1.10.2",
]

[tool.uv]
package = false
